provides a version read-only api v2 and a full hypermedia v1 api for features and accounts
