+++ Rakefile
@@ -3,15 +3,16 @@ require File.expand_path('../config/boot', __FILE__)
 
 Dir.glob('./lib/tasks/*.rake').each { |r| import r }
 
-require 'rspec/core/rake_task'
-require 'rubocop/rake_task'
+if Gris.env.development? || Gris.env.test?
+  require 'rspec/core/rake_task'
+  require 'rubocop/rake_task'
+  RSpec::Core::RakeTask.new
 
-RSpec::Core::RakeTask.new
+  desc 'Run RuboCop'
+  RuboCop::RakeTask.new(:rubocop) do |task|
+    task.fail_on_error = true
+    task.options = %w(-D --auto-correct)
+  end
 
-desc 'Run RuboCop'
-RuboCop::RakeTask.new(:rubocop) do |task|
-  task.fail_on_error = true
-  task.options = %w(-D --auto-correct)
+  task default: [:rubocop, :spec]
 end
-
-task default: [:rubocop, :spec]
