+++ app/api/application_endpoint.rb
@@ -1,0 +1,4 @@
+class ApplicationEndpoint < Grape::API
+  mount V1::Endpoints::RootEndpoint
+  mount V2::Endpoints::RootEndpoint
+end
