+++ app/api/v1/presenters/accounts_presenter.rb
@@ -1,0 +1,9 @@
+module V1
+  module Presenters
+    module AccountsPresenter
+      include Gris::PaginatedPresenter
+
+      collection :to_a, extend: AccountPresenter, as: :accounts, embedded: true
+    end
+  end
+end
