+++ app/api/v1/presenters/features_presenter.rb
@@ -1,0 +1,9 @@
+module V1
+  module Presenters
+    module FeaturesPresenter
+      include Gris::PaginatedPresenter
+
+      collection :to_a, extend: FeaturePresenter, as: :features, embedded: true
+    end
+  end
+end
