--- app/presenters/accounts_presenter.rb
@@ -1,5 +1,0 @@
-module AccountsPresenter
-  include Gris::PaginatedPresenter
-
-  collection :to_a, extend: AccountPresenter, as: :accounts, embedded: true
-end
