--- app/presenters/features_presenter.rb
@@ -1,5 +1,0 @@
-module FeaturesPresenter
-  include Gris::PaginatedPresenter
-
-  collection :to_a, extend: FeaturePresenter, as: :features, embedded: true
-end
