+++ spec/api/cors_spec.rb
